Policy manager overhaul + removed elements following meeting and more
Removed elements from resources following the meeting and added a hidden value that determines how much safety there is against the disasters that planners/es_workers compensate for.
Scrapped public favour and awareness in policies. Instead policies will give a buff to the amount of gain u get for a specific resource. They also have a requirement level for population in a faction and costs you a specific amount of resources.

To do:
Actually selecting a policy and when that happens it discards the card, uses the effects and costs and checks the requirement and then re-add another policy card on the current set.
Possible additions:
Possible nerfs, more than one buff.