Added pip selector variable in fmod
Created a pipDirection labelled param in fmod, the event will trigger a different beeping sound depending on whether the player is adding or removing points.

An addition to this event could be the ability to determine if no points have been assigned, then could have just the spinning dial sound and no points beep.