Tweaked sequence timings
Did some memory management. Due to nature of disasters being able to double trigger, this means if i were to use stealing in fmod, there would be a small change the wrong audio would trigger, at it would steal to second event. Despite releasing the disater one shot immediately unity took a while to release it, trying a bit of memory tweaks in fmod to be able to decrease sequence year lock out